refactor: simplify openCondition handling
