+++ src/app/[eventSlug]/actions.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import type { EventDetailsKey } from "@/i18n/translate-or-fallback";
+import type { EventDetailsKey } from "@/i18n/utils";
 import { API_URL } from "@/lib/api";
 import { isValidUuid } from "@/lib/is-valid-uuid";
 import type { FormErrorObject } from "@/types/form";
