+++ src/app/auth/password-reset/page.tsx
@@ -20,7 +20,7 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { useToast } from "@/hooks/use-toast";
-import { translateOrFallback } from "@/i18n/translate-or-fallback";
+import { translateOrFallback } from "@/i18n/utils";
 import type { ResetError } from "@/types/auth";
 import { resetPasswordSchema } from "@/types/schemas";
 import type { AuthSchemaErrorKeys } from "@/types/schemas";
