+++ src/app/dashboard/(create-event)/actions.ts
@@ -2,7 +2,7 @@
 
 import { formatISO } from "date-fns";
 
-import type { DashboardKey } from "@/i18n/translate-or-fallback";
+import type { DashboardKey } from "@/i18n/utils";
 import { API_URL } from "@/lib/api";
 import { generateFileFromDataUrl } from "@/lib/event";
 import { verifySession } from "@/lib/session";
