+++ src/app/dashboard/events/[uuid]/blocks/[blockUuid]/block-form.tsx
@@ -17,7 +17,7 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { translateOrFallback } from "@/i18n/translate-or-fallback";
+import { translateOrFallback } from "@/i18n/utils";
 
 export const BlockSchema = z.object({
   name: z.string().min(1, "blockNameRequired"),
