+++ src/app/dashboard/events/[uuid]/participants/table/components/buttons/export-button.tsx
@@ -12,7 +12,7 @@ import {
   TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { useToast } from "@/hooks/use-toast";
-import { translateOrFallback } from "@/i18n/translate-or-fallback";
+import { translateOrFallback } from "@/i18n/utils";
 import { downloadFile } from "@/lib/utils";
 
 export function ExportButton({ eventUuid }: { eventUuid: string }) {
