+++ src/components/forms/event/personalization-form.tsx
@@ -16,7 +16,7 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { translateOrFallback } from "@/i18n/translate-or-fallback";
+import { translateOrFallback } from "@/i18n/utils";
 import { cn } from "@/lib/utils";
 
 // Required for usage of useFieldArray hook
